requirements.txt
boto3
celery
dateparser
Django>=5,<6
django-countries
django-localflavor
django-taggit
django-timezone-field
djangorestframework
packaging
pillow
psycopg2-binary
pycurl
pyparsing
pytz
ipython
requests
six
defusedxml
